+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The MiniMax class is no longer a part of the API
 - The DummyBoardEvaluator is no longer a part of the API
 - The MiniMaxAlphaBeta is no longer a part of the API
+- All API methods from MoveStrategy
 
 ### Rename
 - BoardInterface into Board
 
 ### Added
 - Default method providing builder for MoveStrategy
+- One high cohesive method for searching move in MoveStrategy
 
 ## 3.5.0 - 2020.12.12
 ### Removed